changed /fog command to /sfog since /fog is an interal clientside only command; renamed dynamicWeather.ts to weathermanager.ts; converted dynamic weather worker to a class (WeatherManager); created folders 'managers' and 'models' to reorganize classes, enums, interfaces, etc
